# backend/kor/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "KOR API Swagger",
        "title": "KOR API Swagger",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/example-get": {
            "get": {
                "description": "An example GET API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Example GET endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.response"
                        }
                    }
                }
            }
        },
        "/api/v1/example-post": {
            "post": {
                "description": "An example POST API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Example POST endpoint",
                "parameters": [
                    {
                        "description": "Post Request Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.postRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.response"
                        }
                    }
                }
            }
        },
        "/healthcheck": {
            "get": {
                "description": "Returns the status of the server",
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.postRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "main.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}